todo casi listo, a punto de probar poner null el producto en factura

+++ entities/Dicta.js
@@ -3,7 +3,7 @@ import sequelize from '../config/database';
 
 const Dicta = sequelize.define('dicta', {
   idclase:{
-    type: DataTypes.INTEGER,
+    type: DataTypes.INTEGER(9),
     allowNull: false,
     primaryKey: true,
     autoIncrement: true
